Creo clases para un publicador y un comprador

+++ backend/src/main/java/ar/edu/utn/frba/tacs/tp2024c1/grupo1/service/EstadisticasService.java
@@ -2,7 +2,6 @@ package ar.edu.utn.frba.tacs.tp2024c1.grupo1.service;
 
 import ar.edu.utn.frba.tacs.tp2024c1.grupo1.dto.EstadisticaDTO;
 import ar.edu.utn.frba.tacs.tp2024c1.grupo1.repository.EsArticuloRepository;
-import ar.edu.utn.frba.tacs.tp2024c1.grupo1.repository.EsUsuarioRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -13,8 +12,8 @@ public class EstadisticasService {
     private final EsArticuloRepository articuloRepository;
 
     public EstadisticaDTO getEstadisticasBasicas() {
-        int cantUsuariosInteractivos = usuarioService.getTotalUsuariosInteractivos();
-        int cantPublicacionesCreadas = articuloRepository.findAll().size();
+        long cantUsuariosInteractivos = usuarioService.getTotalUsuariosInteractivos();
+        long cantPublicacionesCreadas = articuloRepository.findAll().size();
 
         return new EstadisticaDTO(cantUsuariosInteractivos, cantPublicacionesCreadas);
     }
